tv details refactoring
