+++ lib/ui/widgets/tv_details/components/tv_details_data.dart
@@ -1,5 +1,5 @@
 // Project imports:
-import 'package:comics_db_app/ui/widgets/tv_details/components/created_by_data.dart';
+import 'package:comics_db_app/ui/widgets/tv_details/components/tv_details_created_by_data.dart';
 import 'package:comics_db_app/ui/widgets/tv_details/components/tv_details_actor_data.dart';
 import 'package:comics_db_app/ui/widgets/tv_details/components/tv_details_content_ratings_data.dart';
 import 'package:comics_db_app/ui/widgets/tv_details/components/tv_details_people_data.dart';
