+++ lib/ui/widgets/tv_details/components/tv_details_description_widget.dart
@@ -8,8 +8,8 @@ import 'package:provider/provider.dart';
 import 'package:comics_db_app/ui/components/custom_description_expandable_text_widget.dart';
 import 'package:comics_db_app/ui/widgets/tv_details/tv_details_cubit.dart';
 
-class TvDescriptionWidget extends StatelessWidget {
-  const TvDescriptionWidget({super.key});
+class TvDetailsDescriptionWidget extends StatelessWidget {
+  const TvDetailsDescriptionWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
