+++ lib/ui/widgets/tv_details/components/tv_details_rating_widget.dart
